refactor(shared): relocate amount utils to shared module
- amount.utils.ts: move from features/expenses/utils to shared/utils (pure relocation)
- amount.utils.spec.ts: move alongside implementation to shared/utils
- shared/utils/index.ts: new barrel re-exporting centsToInputEuros, formatEuros, parseEurosToCents
- DeleteExpenseDialog.tsx: update formatEuros import to ~/shared/utils
- ExpenseCard.tsx: update formatEuros import to ~/shared/utils
- ExpenseFormDialog.tsx: update centsToInputEuros import to ~/shared/utils
- ExpensesHeader.tsx: update formatEuros import to ~/shared/utils
- expense.schema.ts: update formatEuros and parseEurosToCents imports to ~/shared/utils
