+++ apps/web/src/shared/utils/index.ts
@@ -1,0 +1,1 @@
+export { centsToInputEuros, formatEuros, parseEurosToCents } from './amount.utils'
